Argument mismatch
In "Add real test" the tests call the "run_{username}ContigFilter_max" function with an addition argument of self.wsName which is an extra argument to those defined in during function def at the beginning of the Implement tutorial:  "def run_{username}ContigFilter_max(self, ctx, params):". 
Upon removing the extra arg the tests ran fine. In addition "self.wsName" is defined in the "params" so there is no reason to pass it as an arg since it's already passing with the params. Thus, no functionality in the app changes.