spelling/sentence fix
The specification file paragraph had a small typo and a kind of nonsense sentence. The typo was "an structure" and the sentence was "to specify which functions the module that will be accessible to other SDK services as well as the input and output types."

Fixed the sentence to "which functions in the module will be accessible to other SDK services, as well as to specify the input and output types."
Thanks!